Switch login to hosts to 'automation' user (#370)
Switch login to hosts to 'automation' user

In order to stop using root login from bridge to hosts we introduce
automation user. It will be the only possibility for bridge to connect
to the host (system admins of course are still able to connect).
Once this is done we can disable root ssh login as required by security
regulations.

Reviewed-by: None <None>
Reviewed-by: kucerakk <kucerakk@gmail.com>