Add zuul-client k8 deployment (#415)
Add zuul-client k8 deployment

We might need to have a standalone zuul client to do some maintenance
(i.e. it is required to clean ZK state while doing update to 4.11 while
zuul itself is down). Since we are moving slowly to having zk in the k8
itself we can only reasonably have zuul client in the same k8 (for case
when zk is not accessible from outside).

Reviewed-by: None <None>
Reviewed-by: OpenTelekomCloud Bot <None>