Fix grafana DS creds management
When we want to get Grafana datasource credentials from vault we were
relying on vault lookup plugins in the inventory. While it is good on
its own it causes getting new vars every time inventory gets evaluated.
In addition if we are not accurate we get user and password not even
matching together.
Instead let us read creds from vault if vault_path is set for the DS.